Update js-picker-overview.md
Resolved issue #444

In step 3.1

Current
Click Generate New Password to create an Application secret. While this value is not necessary for the picker, it must have been created.

Proposed change
Click Generate New Password to create an Application secret. While this value is not necessary for the picker, it must be created.

I am an FTE in USDX in the bay area and want to help contribute to our OSS docs in anyway that I can!

Greg