Fix documentation on listItemUniqueId
Clarify that listItemUniqueId refers to the file ID or folder ID and NOT to the ListItem GUID. It will be what's returned in `_api/web/GetFileByServerRelativePath(decodedurl='/path/to/file')` in the `UniqueId` property, which is different than what's returned in `_api/web/GetFileByServerRelativePath(decodedurl='/path/to/file')/ListItemAllFields` in the `GUID` property. The proper fix should probably be to rename the field to `itemUniqueId` instead of `listItemUniqueId` as "ListItem" means something different in SharePoint.